% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{bernoulli}
\alias{bernoulli}
\title{bernoulli
Randomly Samples a fraction of a SciDB Array and returns a new SciDB Array
*cartesian x,y,z formatting}
\usage{
bernoulli(sdbARRAY, name, fraction)
}
\arguments{
\item{sdbARRAY:}{SciDB type array
name: string -  name of new scidb array}
}
\description{
bernoulli
Randomly Samples a fraction of a SciDB Array and returns a new SciDB Array
*cartesian x,y,z formatting
}
\examples{
A     <- scidb("ARRAY",data.frame=FALSE)
  B     <- bernoulli(sdbARRAY=A, name="sub-DATA", fraction=0.1)
}

