% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/counter.R
\name{counter}
\alias{counter}
\title{counter
Returns the count of an Array filtered at a given radius
*Does not count Zero values}
\usage{
counter(sdbARRAY, r)
}
\arguments{
\item{sdbARRAY:}{SciDB Array
r: distance value with which the array will be filtered and counted}
}
\description{
counter
Returns the count of an Array filtered at a given radius
*Does not count Zero values
}
\examples{
A    <- scidb("DATA",data.frame=FALSE)
  N    <- counter(sdbARRAY=A, r=50.5)
}

