% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{euclidean
Creates a covariance Matrix utilizing the Euclidean distance metric
Given two input SciDB Arrays
*Preferably put the Larger Array first as per Paradigm4 suggestion}
\usage{
euclidean(sdbARRAY1, sdbARRAY2, name)
}
\arguments{
\item{sdbARRAY1:}{SciDB Array
sdbARRAY2: SciDB Array}
}
\description{
euclidean
Creates a covariance Matrix utilizing the Euclidean distance metric
Given two input SciDB Arrays
*Preferably put the Larger Array first as per Paradigm4 suggestion
}
\examples{
A    <- scidb("ARRAY1",data.frame=FALSE)
  B    <- scidb("ARRAY2",data.frame=FALSE)
  C    <- euclidean(sdbARRAY1=A, sdbARRAY2=B, name="Euclidean")
}

