% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ifdel.R
\name{ifdel}
\alias{ifdel}
\title{ifdel
Deletes a given SciDB array if it exists within the database
given an input Array-name from SciDB}
\usage{
ifdel(name)
}
\arguments{
\item{name:}{Name of SciDB Array}
}
\description{
ifdel
Deletes a given SciDB array if it exists within the database
given an input Array-name from SciDB
}
\examples{
name<-"data"
  ifdel(name)
  OUTPUT: "Array Deleted"
}

