% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/make.3D.R
\name{make.3D}
\alias{make.3D}
\title{tpcf
Given an input 1D SciDB Array with Cartesian coordinates
Returns a 3D SciDB Array with dimensions of 1000 times the values
to maintain uniqueness but allow for fast regional querying}
\usage{
make.3D(sdbARRAY, name)
}
\arguments{
\item{sdbARRAY:}{Input 1D scidb array with cartesian coordinates x,y,z
name: Name of Resulting SciDB Array}
}
\description{
tpcf
Given an input 1D SciDB Array with Cartesian coordinates
Returns a 3D SciDB Array with dimensions of 1000 times the values
to maintain uniqueness but allow for fast regional querying
}
\examples{
A     <- scidb("ARRAY",data.frame=FALSE)
  A.3D  <- make.3D(sdbARRAY=A,name="Array3D")
}

