% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/peebles.R
\name{peebles}
\alias{peebles}
\title{peebles
Davis & Peebles: Two-Point correlation Function Estimator (1983)}
\usage{
peebles(DD, DR, ND, NR)
}
\arguments{
\item{DD,DR:}{numeric arrays of DATA-DATA and DATA-RANDOM counts
ND,NR: integer values of the size of the DATA and RANDOM sets respectively}
}
\description{
peebles
Davis & Peebles: Two-Point correlation Function Estimator (1983)
}
\examples{
A     <- scidb("ARRAY",data.frame=FALSE)
  RESULTS <- peebles(DD=DD.counts,DR=DR.counts,ND=100,NR=1000)
}

