% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/poisson.scidb.R
\name{poisson.scidb}
\alias{poisson.scidb}
\title{poisson.scidb
This function allows you to create a random poisson sampling of data points in cartesian space,
given an input Array from SciDB}
\usage{
poisson.scidb(sdbARRAY, name, factor = 1)
}
\arguments{
\item{sdbARRAY:}{SciDB Array
name: string - name of new SciDB array
factor: integer - How much larger than the base Array - DEFAULTS to 1}
}
\description{
poisson.scidb
This function allows you to create a random poisson sampling of data points in cartesian space,
given an input Array from SciDB
}
\examples{
A <- scidb('DATA', data.frame=FALSE)
  R <- poisson.scidb(sdbARRAY=A,factor=10)
}

