% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/target.region.R
\name{target.region}
\alias{target.region}
\title{target.region
Given an input SciDB Array and the minimum and maximum values for x,y,and z coordinates
Allows for efficient analyses of target region}
\usage{
target.region(sdbARRAY, name, xmin, ymin, zmin, xmax, ymax, zmax)
}
\arguments{
\item{sdbARRAY:}{SciDB Array
name: Name of New SciDB Array
xmin,ymin,zmin,xmax,ymax,zmax: double precision coordinate values to bound region}
}
\description{
target.region
Given an input SciDB Array and the minimum and maximum values for x,y,and z coordinates
Allows for efficient analyses of target region
}
\examples{
A    <- scidb("ARRAY1",data.frame=FALSE)
  REG  <- target.region(sdbARRAY=A,name="SubRegion",xmin  = 0.0, ymin  = 0.0,zmin  = 0.0, xmax  = 50.0,ymax  = 50.0, zmax  = 50.0)
}

