% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/thin.R
\name{thin}
\alias{thin}
\title{thin
Progressively Samples a fraction of a SciDB Array and returns a new SciDB Array
Chooses every Nth element of the SciDB Array
*cartesian x,y,z formatting}
\usage{
thin(sdbARRAY, name, factor)
}
\arguments{
\item{sdbARRAY:}{SciDB type array
name: string -  name of new scidb array}
}
\description{
thin
Progressively Samples a fraction of a SciDB Array and returns a new SciDB Array
Chooses every Nth element of the SciDB Array
*cartesian x,y,z formatting
}
\examples{
A     <- scidb("ARRAY",data.frame=FALSE)
  B     <- thin(sdbARRAY=A, name="sub-DATA", factor=100)
}

