% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/tpcf.R
\name{tpcf}
\alias{tpcf}
\title{tpcf
Given three input SciDB Arrays and an array of filtering Radii
Creates a dataframe of counts over each radius value
*suggested input of DATA-DATA,RANDOM-RANDOM,and DATA-RANDOM covariance
distance matrices}
\usage{
tpcf(DD, RR, DR, R, status = TRUE)
}
\arguments{
\item{DD:}{SciDB Array
RR: SciDB Array
DR: SciDB Array
R: numeric array of Radii values
status: Print completion status DEFAULTS to TRUE}
}
\description{
tpcf
Given three input SciDB Arrays and an array of filtering Radii
Creates a dataframe of counts over each radius value
*suggested input of DATA-DATA,RANDOM-RANDOM,and DATA-RANDOM covariance
distance matrices
}
\examples{
DD    <- scidb("ARRAY1",data.frame=FALSE)
  RR    <- scidb("ARRAY2",data.frame=FALSE)
  DR    <- scidb("ARRAY3",data.frame=FALSE)
  R     <- numeric(1.0,5.0,10.0,50.0)
  DATA  <- tpcf(DD=DD, RR=RR, DR=DR, R=R, status=TRUE)
}

